+++ lucaterm/CHANGELOG.md
@@ -1,4 +1,5 @@
 ## LucaTerm master
 
-* handle cursor on journal detail
-* edit journal
+* `luca book`: handle cursor on journal detail
+* `luca book`: edit journal
+* `luca book`: move to another month